refactor: :recycle: Move ~/utils/models to ~/utils/schemas.
