+++ utils/index.ts
@@ -2,4 +2,4 @@ export * from "./requests";
 export * from "./configs";
 export * from "./errors";
 export * from "./strings";
-export * from "./models";
+export * from "./schemas";
